% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinsolventieregistr.R
\name{shinsolventieregistr}
\alias{shinsolventieregistr}
\title{Search for companies in the register in a certain place}
\usage{
shinsolventieregistr(plaatsnaam)
}
\arguments{
\item{plaatsnaam}{Plaatsnaam waarvoor gezocht wordt}
}
\value{
een dataframe met kolommen die niet nested zijn
}
\description{
Look for all the companies in the register based on the name of a city
}
